[imp #1] add support for changes handler
CouchDB exposes an API to be notified whenever a change occurs in the
database.
using the `onChange` methods on a database object, one can register
handlers that are called whenever a change occurs. It accepts the
identifier of the changed document and the optional last version of the
document. The document is empty, if it was deleted

closes #1