first support for design documents
for the moment, one can generate a design document with the DSL if one
respects the following limitation: CommonJS modules are not supported
yet (even if the `require` function is present).

No check is performed on required modules. The dsl code generator should
have a symbol table for libraries to check whether they can be required.
Moreover, restrictions in function uses are not checked either for the
moment