remove legacy server code
writing a server that natively supports scala is not as a good idea as
one may think in the current situation.
The view functions are recompiled for each call which is not optimized
if one wants to compile designs as Scala classes. Using the interpreter
is not optimized either and is probably not suitable for production
environments.
Now that we have a support for a DSL, it is probably better to use it
and compile to javascript instead. More work will be done on the DSL
rather than on the server code "as is".
If the protocol for query servers is improved as proposed here
https://issues.apache.org/jira/browse/COUCHDB-1743 we might reconsider
implementing a server for Scala.