[new] add document patch method
Users that want to work with document patches can now tell sohva to
patch a certain revision of a document and save the patched version into
the database.
If the revision to patch is not the last one, the standard automatic
conflict management process is started to try to solve the conflicts.