[del] remove guessing of attachment mime type
Sohva used to try to guess the mime type of an attachment to upload.
This was done using an external library, adding one more dependency to
the project that was not really needed.
The best way to go about it, is that the person who wishes to upload a
file must say what type it has.
Ask you this: if you don't know what you're doing, then why are you
doing it in the first place ?