database interface cleanup
Version 0.5 introduced the `IdRev` trait that is more efficient than the
structural `Doc` type as it avoids a reflective call. There were some
methods that where duplicated to accept both `Doc` and `IdRev`.
By removing the `Doc` version and creating an implicit view from `Doc`
to `IdRev`, we can now avoid to duplicate code and have a more uniform
interface