Make query to `_all_docs` configurable
`_all_docs` is actually a built-in view for a database. It means that it
accepts all standard view parameters when querying it.
By adding the concept of built-in view directly in sohva, we avoid
reimplementing the same processing for this kind of views as for design
views, and allow for more flexibility for the user.