[imp #14] Choose view types on query
Currently, one has to choose the types of the key, values and documents
upon view creation. This is not well suited for CouchDB, especially for
value types, which may change whether we reduce the view result or not.
By making this choice only on view query, we can have more fine-grained
control over the expected types.
Ideally, it would be even better if, e.g. a query with `include_docs`
set to `false` wouldn't require the user to provide the `Doc` type.

This is the first step toward issue #14.