[imp #17] Add raw view API
This is raw in the sense that objects are not extracted to scala classes
but instead are sent back as json objects. This can be useful when
dealing with views that may return different object types. Use it if you
have no usable (I mean something more specific than `Anyi`) common super
type for your results and/or documents.

This contributes to #14 and close #17