Add the TedMosbyStrategy
This is the contrary of the BarneyStinsonStrategy, and it always keeps
the version from the database in case of a conflict. This also applies
to deleted documents which stay deleted if there is a conflict. This
results in an exception being thrown in such a case because no document
could be saved in the database after all.