Migrate to spray-json
spray-json has shorter release cycles than lift-json and has less magic.
Even if it may appear heavier to use at a first sight, due to the
required `JsonFormat` in scope, it is probably less surprising. Also,
the fact that it is not relying on reflection will make it easier to
port sohva to Scala.js in the future.