[fix #72] Use custom design document format
Most of the fields in a design documents are optional. To ensure that
Sohva can retrieve design document saved by other means in the database,
absence of these fields must be correctly handled. The custom format
treat their abasence as an empty collection and does not write the to
Json if they are empty.